chore: rename to endtoend tests
