+++ tests/endtoend/structsinpkg/type1.go
