+++ tests/endtoend/structsinpkg/type1_validator.go
